requirements.txt
streamlit
tensorflow
Pillow
matplotlib